Added signal handling to Breakdown Manager

+++ src/libs/MsgQueue.h
@@ -7,4 +7,4 @@ typedef struct {
 } msg;
 
 int mqid;
-int msglen;
+size_t msglen;
